return the Promise that was promised (by the docs) ?
This gives an example of returning the Promise (that the docs say `addConnection()` should be returning) ... is this what the intention was? what of the fact that the `Topology` instance is no longer returned? do users need/want to have that instance exposed (or should it actually be hidden as an implementation detail)?