fix: msg.reply serializes to json if body is of type Array
When you send an array in publish or request, it's for the subscriber
not possible to decipher it because it will be considered a buffer
(octet stream).
Writing array structures as json seems a valid pattern but the
serializer would not allow it and limit a user of rabbot to wrap array
structure in an object body.